vps-python-bot/dashboard/requirements.txt
streamlit==1.26.0
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.0.1
oandapyV20==0.7.2
cryptography==41.0.3